check for python instead of jupyter
+ not sensitive to `--user` installs
+ more likely to follow symlinks to sys.prefix (homebrew)
- might catch the wrong Python

It should be hard to catch the wrong Python in a way that matters, though.
The only real case is Jupyter installed with Python 3,
no envs (not conda, not virtualenv, not python brew, etc.),
and `which python` is python2
and python3 has an unusual prefix (not /usr/local, not /usr).

It could also happen where a virtualenv is active,
but jupyter is not in the env, not on /usr:/usr/local,
yet still on PATH.