feat(cell): keep cell in edit mode + cell type shortcut
This commit addresses issue #1360
When a code cell is converted to a markdown cell (and there is content in the cell editor) the editor will be refocused and kept in edit mode.

The commit adds keyboard shortcuts for switching between code and markdown cells.
- When in a code cell "ALT+m" will switch to markdown.
- When in a markdown cell "ALT+c" will switch to a code cell.

Looking for feedback on where I should document the shortcut.