ux(editor): Poking around completer API for custom rendering
Just Poking around to see how to get our custom rendering function.
It's a bit dumb as you need to pass the render function for each
completion item and not to as a global render function.

I guess that could be patched in codemirror upstream, but let's do this
for now.

From the UI-side of things it should not change much, but that should
allow to test new completer UI for the extra information that can be
returned from ipykernel (that needs patches).

documentation and style

Handle metadata unset.