bootstrap a core package
since exporting at a granular level from the desktop application
is proving to feel more constraining (to me) than using our actions
and reducers as a whole, I'm going to try out making a `core` package
here that uses them directly as is here. We can keep this package
free of electron specific bits and export the actions / reducers / epics
as modules to incorporate within both desktop and the web app(s).