watch some events on the zmq sockets
fairly pointless though it seems -- I'm not seeing a way to grab these events

The reason I'm doing this, at least for now is that I was intermittently getting into
a dead kernel situation where the channels were "connected" yet clearly not doing anything.

Hoping to figure out a way to detect scenarios like this and recover.