Model kernel "launching" state using a Record rather than a Map.
Fixes #3417

Previously, reducers in kernel.js reacted to LAUNCH_KERNEL/LAUNCH_KERNEL_BY_NAME
by populating kernels.byRef[newKernelRefUUID] with a Map({state: "launching"}).
Later, when more complete kernel info is available (LAUNCH_KERNEL_SUCCESSFUL),
the Map was replaced with either a LocalKernelRecord or a RemoteKernelRecord.
Reading Maps and Records isn't symmetrical and so this creates issues like #3417.

The fix applied here is to require dispatchers of either launch action to declare
whether the kernel they're launching is local or remote, so that we can create
the proper state record right away. We do not solve the problem of determining
local vs. remote by inspecting a source of truths about kernels; we currently
just assume the desktop app uses local ZMQ kernels and everyone else uses remote
websocket ones. This will need to be revisited if/when that changes.