Dropping images: Holding down the ctrl key also copies the images
to the notebook path.

As far as I know, holding down the alt/options key is the default "copy" modifier key for drag-and-drop operations in macOS. Holding down the ctrl key is the default "copy" modified key for drag-and-drop operations in linux and windows. I did not find a more abstract way to implement this, e.g. with Electron's accelerators.

https://github.com/nteract/nteract/issues/5081