Add feature to insert pasted image files as images in new markdown cells
This early version only works on macOS and requires the `plist` npm package.
https://github.com/nteract/nteract/issues/4963#issuecomment-627561034

The files are not embedded in the ipynb file, but the markdown images just refer to their original file paths. If the image paths are within the notebook directory, relative paths are used to refer to the image files.

https://github.com/nteract/nteract/issues/4963
https://github.com/nteract/nteract/issues/5081